separate LionTest class into parameterized and non-parameterized tests, rename tests
