+++ src/test/java/com/example/CatTest.java
@@ -11,20 +11,20 @@ public class CatTest {
     private Cat cat;
 
     @Before
-    public void setUpTest() {
+    public void setUp() {
         Feline feline = new Feline();
         cat = new Cat(feline);
     }
 
     @Test
-    public void getSoundTest() {
+    public void checkGetSoundReturnsCatSound() {
         String actualResult = cat.getSound();
         String expectedResult = "Мяу";
         assertEquals(expectedResult, actualResult);
     }
 
     @Test
-    public void getFoodTest() throws Exception {
+    public void checkGetFoodReturnsCatFoodList() throws Exception {
         List<String> actualResult = cat.getFood();
         List<String> expectedResult = List.of("Животные", "Птицы", "Рыба");
         assertEquals(expectedResult, actualResult);
