+++ src/test/java/com/example/LionTest.java
@@ -3,38 +3,19 @@ package com.example;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 
-@RunWith(Parameterized.class)
 public class LionTest {
     private Lion lion;
-
-    private final String sex;
-    private final Object expectedResult;
-
-    public LionTest(String sex, Object expectedResult) {
-        this.sex = sex;
-        this.expectedResult = expectedResult;
-    }
-
-    @Parameterized.Parameters
-   public static Object[][] getTestData() {
-       return new Object[][]{
-               {"Самец", true},
-               {"Самка", false},
-       };
-   }
 
     @Before
     public void setUp() {
         Feline feline = new Feline();
         try {
-            lion = new Lion(sex, feline);
+            lion = new Lion("Самец", feline);
         } catch (Exception exception) {
             String textException = "Используйте допустимые значения пола животного - самец или самка";
             Assert.assertEquals(textException, exception.getMessage());
@@ -42,20 +23,14 @@ public class LionTest {
     }
 
     @Test
-    public void getKittensTest() {
+    public void checkGetKittensReturnsValueOne() {
         int actualResult = lion.getKittens();
         int expectedResult = 1;
         assertEquals(expectedResult, actualResult);
     }
 
     @Test
-    public void doesHaveManeTest() {
-        boolean actualResult = lion.doesHaveMane();
-        assertEquals(expectedResult, actualResult);
-    }
-
-    @Test
-    public void doesHaveManeTest1() {
+    public void checkDoesHaveManeThrowsException() {
         Feline feline = new Feline();
         try {
             lion = new Lion("Не опознан", feline);
@@ -66,7 +41,7 @@ public class LionTest {
     }
 
     @Test
-    public void getFoodTest() throws Exception {
+    public void checkGetFoodReturnsLionFoodList() throws Exception {
         List<String> actualResult = lion.getFood();
         List<String> expectedResult = List.of("Животные", "Птицы", "Рыба");
         assertEquals(expectedResult, actualResult);
